% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post.package.R
\name{post.package}
\alias{post.package}
\title{Post Package Setup}
\usage{
post.package()
}
\value{
post.package
}
\description{
This function does the final setup of the newly created package.
No parameters are required.
}
