% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.template.package.creator.R
\docType{package}
\name{utils.template.package.creator}
\alias{utils.template.package.creator}
\alias{utils.template.package.creator-package}
\title{utils.template.package.creator: A package for creating template packages.}
\description{
This package provides two helper functions to create packages incorporating
markdown templates.
}
\details{
\code{\link{make.template.package}}: is run from the R console inside R Studio
and creates the package framework.

\code{\link{post.package}}: is run from inside R Studio after the package has
been loaded.
}
\section{External Requirements}{
 As a minimum, you will need R, RStudio, Pandoc
 and Git correctly installed on your system. Extensive operating dependent
 instructions are avaialble from \url{http://cran.r-project.org}. Please be
 careful of following unofficial instructions as some are just wrong.
}

