% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.template.package.R
\name{make.template.package}
\alias{make.template.package}
\title{Make Template Package}
\usage{
make.template.package(set.my.name, set.package.name)
}
\arguments{
\item{set.my.name}{string with user name}

\item{set.package.name}{string with package name}
}
\value{
make.template.package
}
\description{
Make Template Package
}
